Time: 7 ms (91.84%) | Memory: 6.3 MB (79.90%) - LeetSync

+++ 326-power-of-three/power-of-three.cpp
@@ -4,7 +4,7 @@ public:
         if(n <= 0){
             return false;
         }
-        long long int k=pow(3,19);
-        return ((k) % n == 0);
+        int k=pow(3,19);
+        return (k % n == 0);
     }
 };
